Extract parseAttributes() from utility.ts into a new file and cover this function with tests
